Format lobby screen
