+++ src/App.css
@@ -3,7 +3,7 @@
   width: 315px;
 } */
 .screen {
-  font-family: Arial, sans-serif;
+  /* font-family: Arial, sans-serif; */
   padding-top: 10px;
   padding-bottom: 10px;
 }
