+++ src/components/LobbyRoster/index.css
@@ -1,0 +1,3 @@
+table.lobby-roster {
+  width: 175px;
+}
